fix: quantity_sold as float across schemas and dashboard
- sales_schema.py: quantity_sold int -> float (was rejecting fractional cut sales)
- stock_schema.py: total_quantity_sold int -> float
- dashboard.py: stop truncating fractional quantity_sold to int

+++ backend/app/routes/dashboard.py
@@ -151,7 +151,7 @@ def product_search(
             .filter(Stocks.product_id == product.id)
             .scalar()
         )
-        quantity_sold = int(sold_result or 0)
+        quantity_sold = float(sold_result or 0)
 
         # Initial quantity = available now + what was sold
         initial_quantity = available + quantity_sold
